[core] Factor out xctoolchain plist parsing NFC
